Move url functionality to his own module
